+++ src/collective/awstats_hitcounter/upgrades.py
@@ -1,0 +1,6 @@
+DEFAULT_PROFILE = 'profile-collective.awstats_hitcounter:default'
+
+def upgrade_1100_to_1200(context):
+    print "Upgrading to 0.1.5"
+    context.runImportStepFromProfile(DEFAULT_PROFILE, 'portlets')
+    context.runImportStepFromProfile(DEFAULT_PROFILE, 'cssregistry')